fix(lint): silence gosec false positives and exclude tests from goconst
- Suppress gosec G703 on GITHUB_OUTPUT path provided by the runner
- Replace b%byte(length) with int(b)%length to clear gosec G115
- Exclude _test.go from goconst/gosec/dupl/gocyclo/unparam
- Drop now-redundant nolint directives in test files

+++ random/random.go
@@ -50,7 +50,7 @@ func StringWithCharset(byteLen int, charset Charset) (string, error) {
 	}
 	length := len(charset)
 	for i, b := range bytes {
-		bytes[i] = charset[b%byte(length)]
+		bytes[i] = charset[int(b)%length]
 	}
 
 	return bytesconv.BytesToStr(bytes), nil
